django_project/requirements.txt
akismet==1.1
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
backcall==0.1.0
bleach==3.1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
confusable-homoglyphs==3.2.0
constantly==15.1.0
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-jsonfield==1.4.0
django-ranged-response==0.2.0
django-registration==3.1
django-simple-captcha==0.5.12
-e git://github.com/DOOMer/django-simple-feedback-form@77f166f8f9e228ad8e48730902b96fc69d33400a#egg=django_simple_feedback_form
django-summernote==0.8.11.6
entrypoints==0.3
filelock==3.0.12
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numpy==1.18.2
oauthlib==3.1.0
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyrsistent==0.16.0
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
txaio==20.4.1
urllib3==1.25.9
virtualenv==20.0.17
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
zope.interface==5.1.0
